Data-QA-engineer/pyproject.toml
[project]
name = "data-pipeline"
version = "0.1.0"
description = "Template data pipeline managed by uv"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Chester", email = "" }]
license = { text = "MIT" }
keywords = ["data", "pipeline", "prefect", "uv", "template"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "typer",
  "rich",
  "loguru",
  "pydantic",
  "pyyaml",
  "python-dotenv",
  "prefect",
  "pandas>=2.3.2",
  "numpy>=2.2.6",
  "polars>=1.33.1",
]

[project.optional-dependencies]
dev = [
  "ruff",
  "mypy",
  "pytest",
]

[project.scripts]
# CLI entry point
pipeline = "data_pipeline.utils.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
# Use src layout
package = true

[tool.uv.sources]
# pin optional indexes here if needed

[tool.uv.workspace]
members = []


[tool.hatch.build.targets.wheel]
packages = ["src/data_pipeline"]

[tool.ruff]
line-length = 79
src = ["src"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.10"
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
no_implicit_optional = true
check_untyped_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
pythonpath = ["src"]

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20250915",
]

